Make RecursiveFactorization.jl optional
RecrusiveFactorization.jl brings in the LoopVectorization.jl stack, and so it's not a small dependency. It is a good thing to default to often, but we can make it a bit more optional in order to greatly decrease load times.
