+++ src/factorization.jl
@@ -989,75 +989,6 @@ function SciMLBase.solve!(cache::LinearCache, alg::CHOLMODFactorization; kwargs.
     SciMLBase.build_linear_solution(alg, cache.u, nothing, cache)
 end
 
-## RFLUFactorization
-
-"""
-`RFLUFactorization()`
-
-A fast pure Julia LU-factorization implementation
-using RecursiveFactorization.jl. This is by far the fastest LU-factorization
-implementation, usually outperforming OpenBLAS and MKL for smaller matrices
-(<500x500), but currently optimized only for Base `Array` with `Float32` or `Float64`.
-Additional optimization for complex matrices is in the works.
-"""
-struct RFLUFactorization{P, T} <: AbstractDenseFactorization
-    RFLUFactorization(::Val{P}, ::Val{T}) where {P, T} = new{P, T}()
-end
-
-function RFLUFactorization(; pivot = Val(true), thread = Val(true))
-    RFLUFactorization(pivot, thread)
-end
-
-function init_cacheval(alg::RFLUFactorization, A, b, u, Pl, Pr, maxiters::Int,
-        abstol, reltol, verbose::Bool, assumptions::OperatorAssumptions)
-    ipiv = Vector{LinearAlgebra.BlasInt}(undef, min(size(A)...))
-    ArrayInterface.lu_instance(convert(AbstractMatrix, A)), ipiv
-end
-
-function init_cacheval(alg::RFLUFactorization, A::Matrix{Float64}, b, u, Pl, Pr,
-        maxiters::Int,
-        abstol, reltol, verbose::Bool, assumptions::OperatorAssumptions)
-    ipiv = Vector{LinearAlgebra.BlasInt}(undef, 0)
-    PREALLOCATED_LU, ipiv
-end
-
-function init_cacheval(alg::RFLUFactorization,
-        A::Union{AbstractSparseArray, AbstractSciMLOperator}, b, u, Pl, Pr,
-        maxiters::Int,
-        abstol, reltol, verbose::Bool, assumptions::OperatorAssumptions)
-    nothing, nothing
-end
-
-function init_cacheval(alg::RFLUFactorization,
-        A::Union{Diagonal, SymTridiagonal, Tridiagonal}, b, u, Pl, Pr,
-        maxiters::Int,
-        abstol, reltol, verbose::Bool, assumptions::OperatorAssumptions)
-    nothing, nothing
-end
-
-function SciMLBase.solve!(cache::LinearCache, alg::RFLUFactorization{P, T};
-        kwargs...) where {P, T}
-    A = cache.A
-    A = convert(AbstractMatrix, A)
-    fact, ipiv = @get_cacheval(cache, :RFLUFactorization)
-    if cache.isfresh
-        if length(ipiv) != min(size(A)...)
-            ipiv = Vector{LinearAlgebra.BlasInt}(undef, min(size(A)...))
-        end
-        fact = RecursiveFactorization.lu!(A, ipiv, Val(P), Val(T), check = false)
-        cache.cacheval = (fact, ipiv)
-
-        if !LinearAlgebra.issuccess(fact)
-            return SciMLBase.build_linear_solution(
-                alg, cache.u, nothing, cache; retcode = ReturnCode.Failure)
-        end
-
-        cache.isfresh = false
-    end
-    y = ldiv!(cache.u, @get_cacheval(cache, :RFLUFactorization)[1], cache.b)
-    SciMLBase.build_linear_solution(alg, y, nothing, cache)
-end
-
 ## NormalCholeskyFactorization
 
 """
